data_collection/barcode-dc/requirements.txt
pyudev==0.22.0
evdev==1.3.0
paho-mqtt==1.5.0
